Switch Scene.props to Preferences.props
When a function is called inside a class, like store_props() was doing, it will only run when the class is initialized (IE when the addon is enabled/blender starts).
What this version does is set the same property in the preferences window, as the 3D Viewport.
If the intended result it to store the property inside the scene but only set the defaults to new scenes, then a function can be ran after registering classes, that sets the parameters from prefs() to Scene.

Currently if the user changes a button in ANY scene then save preferences, those preferences will be the default everywhere.
With that proposed version, their defaults will be stored for future scenes and only that scene will be altered.