Adjust code in create_rigui_id.py
Objects with custom properties are treated like dictionaries, by using object.get('custom prop', "else return other or None")
There's also a certain custom property created+used for other parameters in a custom_prop.  It's called ['_RNA_UI']
Extra:
dir() is a function that returns a list of properties for any object.
If ran alone, dir() returns a list of variables available at the current line. If ran as dir(None), a list of parameters for the None value is returned.
It doesn't return everything; for example dir() doesn't show the .get() function but hasattr('get') does see it.