default self.shift, and bpy.context to context
Invoke is only called when clicking a UI button (if the uilayout didn't change the "operator_context") or the function calling the operator specifies 'INVOKE_DEFAULT' (or the other invokes).
So, if invoke is skipped, it would give an error.

uilayout.context_pointer_set and context.copy() allows you to change items inside of the 'context' variable that invoke/execute/poll uses.
If you specify bpy.context, this ignores those changes, which would make things not work as they intended.
And if they DONT make those kind of changes, it'll still behave as though you did specify bpy.context