Toggle for RNA Properties, as well as proper display for them
When an api-defined item is a PropertyGroup, it's written to the custom properties as a dict().
Before, these would display as an empty box.

The code default Blender UI uses, tries to convert the property to a dict() and display that as text.

It fails to do so for certain property setups, and just displays it as "{}" but it's better than nothing.